Add service_enable and service_ensure parameters
This patch removes the mostly useless manage_service parameter and
instead allows service_ensure and service_enable to be passed to the
galera::server class. The problem with manage_service was that it
could not be set to false, else the galera service would not be
defined. In addition, it was not possible to have the service disabled
yet have ensure set to 'running'. This patch addresses both of these
problem. The service_ensure parameter will default to 'running' since
we need the service running in order to set the root password. The
service_enable parameter will default to 'true', such that the service
is enabled by default.